feat(application) : refactor namespaces and introduce service structure
- Migrated model classes from `CoreXCrud.Models` to `CoreXCrud.Entities`.
- Updated controller files to reflect the new namespace structure.
- Added a new folder for `Models` in the project file.
- Introduced empty service interfaces and classes for `OrderDetail`, `Order`, `Product`, and `User`, organized under their respective namespaces.
- Maintained existing FluentValidation and Serilog integrations for validation and logging.
- This refactor aims to enhance project organization and modularity.

<-- This commit message was generated with GitHub Copilot in Visual Studio. -->
