+++ Entities/Order.cs
@@ -1,7 +1,7 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace CoreXCrud.Models
+namespace CoreXCrud.Entities
 {
     /// <summary>
     /// Sipariş tablosu modeli.
