+++ Repositories/IUnitOfWork.cs
@@ -1,4 +1,4 @@
-﻿using CoreXCrud.Models;
+﻿using CoreXCrud.Entities;
 
 namespace CoreXCrud.Repositories
 {
