+++ Services/OrderDetailServices/OrderDetailService.cs
@@ -1,0 +1,6 @@
+﻿namespace CoreXCrud.Services.OrderDetailServices
+{
+    public class OrderDetailService
+    {
+    }
+}
