+++ Services/OrderServices/OrderService.cs
@@ -1,0 +1,6 @@
+﻿namespace CoreXCrud.Services.OrderServices
+{
+    public class OrderService
+    {
+    }
+}
