+++ Services/ProductServices/IProductService.cs
@@ -1,0 +1,6 @@
+﻿namespace CoreXCrud.Services.ProductServices
+{
+    public interface IProductService
+    {
+    }
+}
