+++ Services/ProductServices/ProductService.cs
@@ -1,0 +1,6 @@
+﻿namespace CoreXCrud.Services.ProductServices
+{
+    public class ProductService
+    {
+    }
+}
