+++ Services/UserServices/IUserService.cs
@@ -1,0 +1,6 @@
+﻿namespace CoreXCrud.Services.UserServices
+{
+    public interface IUserService
+    {
+    }
+}
